+++ client/components/notes/addEditModal.vue
@@ -75,7 +75,7 @@
     //------------------------------------------------------------------------------------------------------------------
 
     // Managers
-    import notesMan from '../../api/managers/notes';
+    import notesMan from '../../api/managers/notebook';
 
     //------------------------------------------------------------------------------------------------------------------
