+++ client/components/notes/deleteModal.vue
@@ -49,7 +49,7 @@
     //------------------------------------------------------------------------------------------------------------------
 
     // Managers
-    import notesMan from '../../api/managers/notes';
+    import notesMan from '../../api/managers/notebook';
 
     //------------------------------------------------------------------------------------------------------------------
